fixed structure of tests and classes
