+++ requirements.txt
@@ -1,0 +1,17 @@
+atomicwrites==1.4.0
+attrs==20.3.0
+certifi==2020.12.5
+chardet==4.0.0
+colorama==0.4.4
+demjson==2.2.4
+idna==2.10
+more-itertools==8.6.0
+numpy==1.19.4
+packaging==20.8
+pluggy==0.13.1
+py==1.10.0
+pyparsing==2.4.7
+pytest==5.1.1
+requests==2.25.1
+urllib3==1.25.10
+wcwidth==0.2.5
